restructure ui header and source files
