+++ include/createHeaderBar.h
@@ -1,0 +1,4 @@
+#ifndef CREATEHEADERBAR_H
+#define CREATEHEADERBAR_H
+#include "../src/ui/createHeaderBar.c"
+#endif
