+++ include/onSsdpScanCompleted.h
@@ -1,0 +1,4 @@
+#ifndef ONSSDPSCANCOMPLETED_H
+#define ONSSDPSCANCOMPLETED_H
+#include "../src/ui/onSsdpScanCompleted.c"
+#endif
