+++ include/runSsdpScan.h
@@ -1,0 +1,4 @@
+#ifndef RUNSSDPSCAN_H
+#define RUNSSDPSCAN_H
+#include "../src/ui/runSsdpScan.c"
+#endif
