+++ include/updateUiOnScan.h
@@ -1,0 +1,4 @@
+#ifndef UPDATEUIONSCAN_H
+#define UPDATEUIONSCAN_H
+#include "../src/ui/updateUiOnScan.c"
+#endif
