+++ src/ui/onStartClicked.c
@@ -1,77 +1,10 @@
-#include "SSDPListenerConnection.h"
-#include "createButton.h"
 #include "createVerticalBox.h"
 #include "createWrappedLabel.h"
-#include "getMyIp.h"
-#include "glib-object.h"
-#include "glib.h"
-#include "gtk/gtk.h"
 #include "joinSSDP.h"
+#include "onSsdpScanCompleted.h"
 #include "removeOldChild.h"
+#include "runSsdpScan.h"
 #include "uiCustomData.h"
-#include "whileDeviceClicked.h"
-
-static void update_ui(gpointer user_data) {
-  UiUpdateData *data = (UiUpdateData *)user_data;
-  g_print("%s", data->result.message);
-
-  // create_sidebar(box, main_content, items, sizeof(items) / sizeof(items[0]));
-  // create_sidebar(data->h_box, data->v_box, data->result.arr,
-  // data->result.size);
-
-  gtk_spinner_stop(GTK_SPINNER(data->spinner));
-  gtk_widget_unparent(data->spinner);
-  gtk_widget_unparent(data->spin_label);
-
-  gtk_widget_set_margin_bottom(data->label_horizontal_box, 0);
-
-  GtkWidget *frame = gtk_frame_new("Found Devices");
-
-  GtkWidget *list_of_device = gtk_list_box_new();
-  gtk_frame_set_child(GTK_FRAME(frame), list_of_device);
-
-  // for demo
-  const char *items[] = {"Apple", "Ball", "Cat"};
-
-  for (int i = 0; i < data->result.size; i++) {
-    GtkWidget *button = create_button(data->result.arr[i],
-                                      G_CALLBACK(while_device_clicked), NULL);
-    gtk_list_box_append(GTK_LIST_BOX(list_of_device), button);
-  }
-  gtk_box_append(GTK_BOX(data->v_box), frame);
-
-  for (int i = 0; i < data->result.size; i++) {
-    g_free(data->result.arr[i]);
-  }
-  g_free(data->result.arr);
-  g_free(data);
-}
-
-static void run_ssdp_scan(GTask *task, gpointer source_object, gpointer data,
-                          GCancellable *cancellable) {
-  g_print("here is run ssdp scan\n");
-  UiUpdateData *ui_data = (UiUpdateData *)data;
-
-  struct ssdpMessage *revMsg = SSDPListen(5);
-  g_print("run ssdp listen function\n");
-  if (g_task_return_error_if_cancelled(task)) {
-    return;
-  }
-
-  g_task_return_pointer(task, revMsg, g_free);
-  g_print("Result of ssdp listen is returned using g_task_return_pointer\n");
-  // ui_data->result = *revMsg;
-  // g_idle_add(update_ui, ui_data);
-}
-
-static void on_ssdp_scan_completed(GObject *source_object, GAsyncResult *res,
-                                   gpointer user_data) {
-  GTask *task = G_TASK(res);
-  UiUpdateData *ui_data = (UiUpdateData *)user_data;
-  struct ssdpMessage *result = g_task_propagate_pointer(task, NULL);
-  ui_data->result = *result;
-  update_ui(ui_data);
-}
 
 void on_start_clicked(GtkButton *button, gpointer data) {
   StartData *user_data = (StartData *)data;
@@ -122,7 +55,6 @@ void on_start_clicked(GtkButton *button, gpointer data) {
   g_task_run_in_thread(task, run_ssdp_scan);
   g_object_unref(task);
 
-  // JoinSSDPThread *thread_data;
   strcpy(thread_data->myIp, findMyIP().message);
   thread_data->timer = 5.0;
   GThread *join_ssdp_thread = g_thread_new("join_ssdp", joinSSDP, thread_data);
